requirements.txt
Django==1.7.1
django-tastypie==0.12.1
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.4.3
six==1.8.0
wsgiref==0.1.2
